backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
requests==2.31.0
python-decouple==3.8
django-extensions==3.2.3
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
dj-database-url==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
djangorestframework-simplejwt==5.3.0